now in a track is a imageurl saved too and little improvement in the speech class
