+++ plugins/spotify/src/main/java/de/unistuttgart/iaas/amyassist/amy/plugin/spotify/logic/PlayerLogic.java
@@ -289,6 +289,7 @@ public class PlayerLogic {
 			case "mute":
 				return setVolume(VOLUME_MUTE_VALUE);
 			case "max":
+			case "full":
 				return setVolume(VOLUME_MAX_VALUE);
 			case "up":
 				volume = Math.min(VOLUME_MAX_VALUE, volume + VOLUME_UPDOWN_VALUE);
